Refactored and reformat code

+++ core/periodic.h
@@ -76,7 +76,7 @@
 #define PERIODIC_TTICKS_PER_SEC (PERIODIC_TOP * CONF_MTICKS_PER_SEC)
 
 /* Macro to convert milliseconds to periodic milliticks */
-#define PERIODIC_MS2MTICKS(msecs) ((msecs) * CONF_MTICKS_PER_SEC / 1000)
+#define PERIODIC_MS2MTICKS(msecs) ((long)(msecs) * CONF_MTICKS_PER_SEC / 1000)
 
 /* Periodic milliticks usability check -
  * there should be at least ~1000 CPU-Cycles per periodic millitick
